requirements.txt
whitenoise==5.2.0
Django==3.1.7
django-cleanup==5.1.0
django-crispy-forms==1.11.2
django-embed-video==1.4.0
requests==2.22.0
Pillow==7.1.2
